refact : 투표 공통부분의 아이디로 일반/전통주 투표 컨텐츠 조회 로직 리팩토링
